cleanup runtime vs. world
