+++ src/lib.rs
@@ -43,22 +43,8 @@ pub mod system;
 /// Tremor runtime version tools
 pub mod version;
 
-use crate::errors::{Error, Result};
-
-use log::{error, info};
-use std::io::Read;
 use system::Runtime;
-use tokio::io::AsyncRead;
-use tremor_script::{
-    ast::{
-        optimizer::Optimizer, visitors::ArgsRewriter, walkers::DeployWalker as _, BaseExpr,
-        CreationalWith, DeployFlow, Helper, Ident, ImutExpr, NodeId, WithExprs,
-    },
-    deploy::Deploy,
-    highlighter::{Dumb as ToStringHighlighter, Highlighter, Term as TermHighlighter},
-    prelude::ValueAsContainer,
-    NodeMeta, FN_REGISTRY,
-};
+use tremor_script::highlighter::Dumb as ToStringHighlighter;
 
 /// Operator Config
 pub type OpConfig = tremor_value::Value<'static>;
@@ -93,115 +79,6 @@ pub(crate) async fn register_builtin_connector_types(runtime: &Runtime) -> anyho
     Ok(())
 }
 
-/// Loads a Troy file
-///
-/// # Errors
-/// Fails if the file can not be loaded
-pub async fn load_troy_file(world: &Runtime, file_name: &str) -> Result<usize> {
-    info!("Loading troy from {file_name}");
-
-    let mut file = tremor_common::file::open(&file_name)?;
-    let mut src = String::new();
-
-    file.read_to_string(&mut src)
-        .map_err(|e| Error::from(format!("Could not open file {file_name} => {e}")))?;
-    let aggr_reg = tremor_script::registry::aggr();
-
-    let deployable = Deploy::parse(&src, &*FN_REGISTRY.read()?, &aggr_reg);
-    let mut h = TermHighlighter::stderr();
-    let deployable = match deployable {
-        Ok(deployable) => {
-            deployable.format_warnings_with(&mut h)?;
-            deployable
-        }
-        Err(e) => {
-            log_error!(h.format_error(&e), "Error: {e}");
-
-            return Err(format!("failed to load troy file: {file_name}").into());
-        }
-    };
-
-    let mut count = 0;
-    for flow in deployable.iter_flows() {
-        world.start_flow(flow).await?;
-        count += 1;
-    }
-    Ok(count)
-}
-
-/// Loads a tremor archive
-///
-/// # Errors
-/// Fails if the file can not be loaded
-pub async fn load_archive(
-    world: &Runtime,
-    archive: impl AsyncRead + Send + Unpin,
-    flow_name: Option<&str>,
-    config: Option<tremor_value::Value<'static>>,
-) -> Result<usize> {
-    let (app, deployable, _indexes) = tremor_archive::extract(archive).await?;
-    info!("App laoded {}", app.name());
-
-    let flow_name = flow_name.unwrap_or(app.entrypoint.as_str());
-
-    // ensure we print the warnings
-    let mut h = TermHighlighter::stderr();
-    deployable.format_warnings_with(&mut h)?;
-
-    // create the config if any is needed
-    let mut with_exprs = Vec::new();
-
-    if let Some(config) = config.as_object() {
-        for (key, value) in config {
-            let value = ImutExpr::literal(NodeMeta::dummy(), value.clone());
-            let ident = Ident::new(key.clone(), NodeMeta::dummy());
-            with_exprs.push((ident, value));
-        }
-    }
-    let with = WithExprs(with_exprs);
-    let with = CreationalWith {
-        with,
-        mid: NodeMeta::dummy(),
-    };
-
-    // since archives are modules they don't have a deoploy statement so we need to find the
-    // flow in the archive create the corresponding deploy statement and start it
-    let mut defn = deployable
-        .deploy
-        .scope
-        .content
-        .flows
-        .get(flow_name)
-        .ok_or_else(|| format!("failed to load archive flow {flow_name}"))?
-        .clone();
-    // ensure we applu the configuration
-    defn.params.ingest_creational_with(&with)?;
-
-    let reg = tremor_script::registry();
-    let aggr_reg = tremor_script::aggr_registry();
-    let mut helper = Helper::new(&reg, &aggr_reg);
-
-    Optimizer::new(&helper).walk_flow_definition(&mut defn)?;
-
-    let inner_args = defn.params.render()?;
-
-    ArgsRewriter::new(inner_args, &mut helper, defn.params.meta())
-        .walk_flow_definition(&mut defn)?;
-
-    Optimizer::new(&helper).walk_flow_definition(&mut defn)?;
-
-    let flow = DeployFlow {
-        mid: NodeMeta::dummy(),
-        from_target: NodeId::new(flow_name, vec![], NodeMeta::dummy()),
-        instance_alias: flow_name.into(),
-        defn,
-        docs: None,
-    };
-
-    world.start_flow(&flow).await?;
-    Ok(1)
-}
-
 /// Logs but ignores an error
 #[macro_export]
 #[doc(hidden)]
@@ -224,8 +101,8 @@ mod tests {
     use std::io::Write;
 
     #[tokio::test(flavor = "multi_thread")]
-    async fn test_load_troy_file() -> Result<()> {
-        let (world, handle) = Runtime::builder()
+    async fn test_load_troy_file() -> crate::errors::Result<()> {
+        let (runtime, handle) = Runtime::builder()
             .default_include_connectors()
             .build()
             .await?;
@@ -244,9 +121,9 @@ mod tests {
         )?;
         troy_file.as_file().flush()?;
         let path = troy_file.path().display().to_string();
-        let num_deploys = load_troy_file(&world, &path).await?;
+        let num_deploys = runtime.load_troy_file(&path).await?;
         assert_eq!(0, num_deploys);
-        world.stop(ShutdownMode::Graceful).await?;
+        runtime.stop(ShutdownMode::Graceful).await?;
         handle.abort();
         troy_file.close()?;
         Ok(())
