+++ src/system.rs
@@ -17,12 +17,26 @@ pub mod flow;
 /// contains the runtime actor starting and maintaining flows
 pub mod flow_supervisor;
 
+use std::io::Read as _;
+
 use self::flow::Flow;
-use crate::errors::{Error, Kind as ErrorKind, Result};
-use log::error;
-use tokio::{sync::oneshot, task::JoinHandle};
+use crate::{
+    errors::{Error, Kind as ErrorKind, Result},
+    log_error,
+};
+use log::{error, info};
+use simd_json::base::ValueAsContainer as _;
+use tokio::{io::AsyncRead, sync::oneshot, task::JoinHandle};
 use tremor_connectors::ConnectorBuilder;
-use tremor_script::{ast, highlighter::Highlighter};
+use tremor_script::{
+    ast::{
+        self, optimizer::Optimizer, visitors::ArgsRewriter, walkers::DeployWalker as _,
+        BaseExpr as _, CreationalWith, DeployFlow, Helper, Ident, ImutExpr, NodeId, WithExprs,
+    },
+    deploy::Deploy,
+    highlighter::{Highlighter, Term},
+    NodeMeta, FN_REGISTRY,
+};
 use tremor_system::{
     killswitch::{KillSwitch, ShutdownMode},
     selector::{PluginType, Rules, RulesBuilder},
@@ -133,6 +147,115 @@ pub struct Runtime {
 }
 
 impl Runtime {
+    /// Loads a Troy file
+    ///
+    /// # Errors
+    /// Fails if the file can not be loaded
+    pub async fn load_troy_file(&self, file_name: &str) -> Result<usize> {
+        info!("Loading troy from {file_name}");
+
+        let mut file = tremor_common::file::open(&file_name)?;
+        let mut src = String::new();
+
+        file.read_to_string(&mut src)
+            .map_err(|e| Error::from(format!("Could not open file {file_name} => {e}")))?;
+        let aggr_reg = tremor_script::registry::aggr();
+
+        let deployable = Deploy::parse(&src, &*FN_REGISTRY.read()?, &aggr_reg);
+        let mut h = Term::stderr();
+        let deployable = match deployable {
+            Ok(deployable) => {
+                deployable.format_warnings_with(&mut h)?;
+                deployable
+            }
+            Err(e) => {
+                log_error!(h.format_error(&e), "Error: {e}");
+
+                return Err(format!("failed to load troy file: {file_name}").into());
+            }
+        };
+
+        let mut count = 0;
+        for flow in deployable.iter_flows() {
+            self.start_flow(flow).await?;
+            count += 1;
+        }
+        Ok(count)
+    }
+
+    /// Loads a tremor archive
+    ///
+    /// # Errors
+    /// Fails if the file can not be loaded
+    pub async fn load_archive(
+        &self,
+        archive: impl AsyncRead + Send + Unpin,
+        flow_name: Option<&str>,
+        config: Option<tremor_value::Value<'static>>,
+    ) -> Result<usize> {
+        let (app, deployable, _indexes) = tremor_archive::extract(archive).await?;
+        info!("App laoded {}", app.name());
+
+        let flow_name = flow_name.unwrap_or(app.entrypoint.as_str());
+
+        // ensure we print the warnings
+        let mut h = Term::stderr();
+        deployable.format_warnings_with(&mut h)?;
+
+        // create the config if any is needed
+        let mut with_exprs = Vec::new();
+
+        if let Some(config) = config.as_object() {
+            for (key, value) in config {
+                let value = ImutExpr::literal(NodeMeta::dummy(), value.clone());
+                let ident = Ident::new(key.clone(), NodeMeta::dummy());
+                with_exprs.push((ident, value));
+            }
+        }
+        let with = WithExprs(with_exprs);
+        let with = CreationalWith {
+            with,
+            mid: NodeMeta::dummy(),
+        };
+
+        // since archives are modules they don't have a deoploy statement so we need to find the
+        // flow in the archive create the corresponding deploy statement and start it
+        let mut defn = deployable
+            .deploy
+            .scope
+            .content
+            .flows
+            .get(flow_name)
+            .ok_or_else(|| format!("failed to load archive flow {flow_name}"))?
+            .clone();
+        // ensure we applu the configuration
+        defn.params.ingest_creational_with(&with)?;
+
+        let reg = tremor_script::registry();
+        let aggr_reg = tremor_script::aggr_registry();
+        let mut helper = Helper::new(&reg, &aggr_reg);
+
+        Optimizer::new(&helper).walk_flow_definition(&mut defn)?;
+
+        let inner_args = defn.params.render()?;
+
+        ArgsRewriter::new(inner_args, &mut helper, defn.params.meta())
+            .walk_flow_definition(&mut defn)?;
+
+        Optimizer::new(&helper).walk_flow_definition(&mut defn)?;
+
+        let flow = DeployFlow {
+            mid: NodeMeta::dummy(),
+            from_target: NodeId::new(flow_name, vec![], NodeMeta::dummy()),
+            instance_alias: flow_name.into(),
+            defn,
+            docs: None,
+        };
+
+        self.start_flow(&flow).await?;
+        Ok(1)
+    }
+
     /// creates a runtime builder
     #[must_use]
     pub fn builder() -> RuntimeBuilder {
